fix: circular import caused zod schemas to break
