+++ .changeset/shiny-news-kneel.md
@@ -1,0 +1,6 @@
+---
+"@assistant-ui/react-ai-sdk": patch
+"@assistant-ui/react-edge": patch
+---
+
+fix: circular import causes zod schemas to break
