+++ packages/react-edge/src/edge/EdgeRuntimeRequestOptions.ts
@@ -1,90 +1,10 @@
-import { JSONSchema7 } from "json-schema";
 import { z } from "zod";
 import {
   LanguageModelConfigSchema,
   LanguageModelV1CallSettingsSchema,
-} from "./createEdgeRuntimeAPI";
-
-const LanguageModelV1FunctionToolSchema = z.object({
-  type: z.literal("function"),
-  name: z.string(),
-  description: z.string().optional(),
-  parameters: z.custom<JSONSchema7>(
-    (val) => typeof val === "object" && val !== null,
-  ),
-});
-
-const TextContentPartSchema = z.object({
-  type: z.literal("text"),
-  text: z.string(),
-});
-
-const ImageContentPartSchema = z.object({
-  type: z.literal("image"),
-  image: z.string(),
-});
-
-const FileContentPartSchema = z.object({
-  type: z.literal("file"),
-  data: z.string(),
-  mimeType: z.string(),
-});
-
-const Unstable_AudioContentPart = z.object({
-  type: z.literal("audio"),
-  audio: z.object({
-    data: z.string(),
-    format: z.union([z.literal("mp3"), z.literal("wav")]),
-  }),
-});
-
-const CoreToolCallContentPartSchema = z.object({
-  type: z.literal("tool-call"),
-  toolCallId: z.string(),
-  toolName: z.string(),
-  args: z.record(z.unknown()),
-  result: z.unknown().optional(),
-  isError: z.boolean().optional(),
-});
-
-const CoreUserMessageSchema = z.object({
-  role: z.literal("user"),
-  content: z
-    .array(
-      z.discriminatedUnion("type", [
-        TextContentPartSchema,
-        ImageContentPartSchema,
-        FileContentPartSchema,
-        Unstable_AudioContentPart,
-      ]),
-    )
-    .min(1)
-    .readonly(),
-});
-
-const CoreAssistantMessageSchema = z.object({
-  role: z.literal("assistant"),
-  content: z
-    .array(
-      z.discriminatedUnion("type", [
-        TextContentPartSchema,
-        CoreToolCallContentPartSchema,
-      ]),
-    )
-    .min(1)
-    .readonly(),
-});
-
-const CoreSystemMessageSchema = z.object({
-  role: z.literal("system"),
-  content: z.tuple([TextContentPartSchema]).readonly(),
-});
-
-const CoreMessageSchema = z.discriminatedUnion("role", [
-  CoreSystemMessageSchema,
-  CoreUserMessageSchema,
-  CoreAssistantMessageSchema,
-]);
+  LanguageModelV1FunctionToolSchema,
+  CoreMessageSchema,
+} from "./schemas";
 
 export const EdgeRuntimeRequestOptionsSchema = z
   .object({
